Updated place_input container
