+++ src/client/scripts/app.jsx
@@ -6,16 +6,16 @@ import { createStore } from 'redux';
 
 import reducer from './reducers/combine_reducers';
 
-import Home from './components/home/home';
-import Planner from './components/planner/planner';
+import Place from './components/place_input';
+import Canvas from './components/canvas';
 
 const store = createStore(reducer);
 
 render((
   <Provider store={store}>
     <Router history={browserHistory}>
-      <Route path="/" component={Home} />
-      <Route path="/planner" component={Planner} />
+      <Route path="/" component={Place} />
+      <Route path="/city" component={Canvas} />
     </Router>
   </Provider>
 ), document.getElementById('container'));
