--- src/client/scripts/components/place/place.scss
@@ -1,11 +1,0 @@
-.red {
-  color: red;
-}
-
-.blue {
-  color: blue;
-}
-
-.orange {
-  color: $brand-danger;
-}
